requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
pandas==2.2.0
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
python-dateutil==2.8.2
Werkzeug==3.0.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
Flask-Login==0.6.3
openpyxl==3.1.2
gunicorn==21.2.0
numpy==1.26.3
pytest==7.4.4
coverage==7.4.1
supervisor==4.2.5
redis==5.0.1
Flask-Session==0.6.0
Flask-Caching==2.1.0 